- added support for direct printing of pdfs at the POS. - added a login blockage for POS-office - changed the way yes/no selection is shown. - fixed problem showing event list for Ticket takers - fixed issue in the installer where a new db was not correctly detected.
git-svn-id: http://repos.svn.fusionticket.org/fusionticket/trunk@1586 87ea6b0e-88cb-4cdd-95e9-9c88242927a5

+++ includes/classes/model.event.php
@@ -283,6 +283,11 @@ class Event Extends Model {
         return self::_abort('discount_delete_failed');
       }
 
+      $query="DELETE FROM adminlink
+              WHERE adminlink_event_id="._esc($this->id);
+      if(!ShopDB::query($query)){
+        return self::_abort('adminlink_delete_failed');
+      }
       $query="DELETE FROM Event
               WHERE event_id="._esc($this->id);
       if(!ShopDB::query($query)){
